Add performance results

+++ profile_solutions.py
@@ -4,7 +4,7 @@ import compilation
 import testing
 import csv
 
-def test_individual_assembly(driver_object_path, assembly_path, test_data_path):
+def test_individual_assembly(driver_object_path, assembly_path, test_data_path, iterations=None):
 	success, compiler_error, compilation_unit_path = compilation.compile_source(assembly_path)
 	if not success:
 		print(f"Compilation failed: {compiler_error}")
@@ -28,7 +28,7 @@ def test_individual_assembly(driver_object_path, assembly_path, test_data_path):
 	# Set executable permissions
 	os.chmod(executable_path, 0o755)
 	
-	success, testing_error, total_cpu_time = testing.run_test_from_csv(test_data_path, executable_path, 10000000)
+	success, testing_error, total_cpu_time = testing.run_test_from_csv(test_data_path, executable_path, iterations)
 	if not success:
 		print("Testing failed:", testing_error)
 		return False, 0
@@ -58,9 +58,26 @@ def handle_problem_directory(problem_directory_path, test_driver_source_path):
 	testDataPath = os.path.join(problem_directory_path, "test_data.csv")
 		
 	generatedDirectoryPath = os.path.join(problem_directory_path, "generated")
+	if not os.path.exists(generatedDirectoryPath):
+		print("Solutions have not been generated.")
+		return
 
 	# List all files in the generated directory and filter for .asm files
 	asm_files = [f for f in os.listdir(generatedDirectoryPath) if f.endswith('.asm')]
+
+	# Determine how many iterations to run
+	unoptimizedAssemblyPath = os.path.join(generatedDirectoryPath, "clang_generated_unoptimized.asm")
+	if not os.path.exists(unoptimizedAssemblyPath):
+		return
+		
+	iterations = 1
+	cpu_time = 0
+	
+	print("Determining iterations…")
+	while cpu_time < 0.2:
+		iterations *= 10
+		success, cpu_time = test_individual_assembly(driverObjectPath, unoptimizedAssemblyPath, testDataPath, iterations)
+	print(f"Using {iterations} iterations.")
 
 	# Create or open a CSV file for writing the results
 	with open(results_csv_path, 'w', newline='') as csvfile:
@@ -71,7 +88,7 @@ def handle_problem_directory(problem_directory_path, test_driver_source_path):
 			fullAssemblyPath = os.path.join(generatedDirectoryPath, filename)
 			
 			# test_individual_assembly returns a tuple (success, cpu_time)
-			success, cpu_time = test_individual_assembly(driverObjectPath, fullAssemblyPath, testDataPath)
+			success, cpu_time = test_individual_assembly(driverObjectPath, fullAssemblyPath, testDataPath, iterations)
 			
 			if success:
 				# Write the results to the CSV file
